requirements.txt
fastapi
uvicorn
finnhub-python
yfinance
python-dotenv
